# storage/app/scribe/openapi.yaml
openapi: 3.0.3
info:
  title: 'portfolio-api API Documentation'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'Use "Authorization: Bearer &lt; JWT &gt;"'
security:
  -
    default: []
paths:
...
  '/api/v1/navbar-items/{navbarItem}':
    patch:
      summary: 'Update navbar item'
      operationId: updateNavbarItem
      description: 'Updates the specified navbar item.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: true
                  message: 'Navbar Item updated successfully.'
                  data:
                    type: NavbarItems
                    id: 3
                    label: Projects
                    link: '/#projects'
                    order: 3
                    links:
                      self: /api/v1/navbar-items/3
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Navbar Item updated successfully.'
                  data:
                    type: object
                    properties:
                      type:
                        type: string
                        example: NavbarItems
                      id:
                        type: integer
                        example: 3
                      label:
                        type: string
                        example: Projects
                      link:
                        type: string
                        example: '/#projects'
                      order:
                        type: integer
                        example: 3
                      links:
                        type: object
                        properties:
                          self:
                            type: string
                            example: /api/v1/navbar-items/3
      tags:
        - 'Navbar Items'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
                  description: 'optional Visible text.'
                  example: Projects
                  nullable: false
                link:
                  type: string
                  description: 'optional URL or hash.'
                  example: '/#projects'
                  nullable: false
                order:
                  type: integer
                  description: 'optional Display order.'
                  example: 3
                  nullable: false
    delete:
      summary: 'Delete navbar item'
      operationId: deleteNavbarItem
      description: 'Removes a navbar item by ID.'
      parameters: []
      responses:
        204:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example: {  }
                properties: {  }
      tags:
        - 'Navbar Items'
    parameters:
      -
        in: path
        name: navbarItem
        description: ''
        example: consequatur
        required: true
        schema:
          type: string
      -
        in: path
        name: id
        description: 'Navbar item ID.'
        example: 3
        required: true
        schema:
          type: integer
  '/api/v1/navbar-items/{id}':
    get:
      summary: 'Get one navbar item'
      operationId: getOneNavbarItem
      description: 'Returns a navbar item by ID.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: true
                  message: OK
                  data:
                    type: NavbarItems
                    id: 2
                    label: About
                    link: '/#about'
                    order: 2
                    links:
                      self: /api/v1/navbar-items/2
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: OK
                  data:
                    type: object
                    properties:
                      type:
                        type: string
                        example: NavbarItems
                      id:
                        type: integer
                        example: 2
                      label:
                        type: string
                        example: About
                      link:
                        type: string
                        example: '/#about'
                      order:
                        type: integer
                        example: 2
                      links:
                        type: object
                        properties:
                          self:
                            type: string
                            example: /api/v1/navbar-items/2
      tags:
        - 'Navbar Items'
    parameters:
      -
        in: path
        name: id
        description: 'Navbar item ID.'
        example: 2
        required: true
        schema:
          type: integer
...